not only fact
remove fact axiom and close fact axiom, since the valuation should not be limit to theorem of a system, but should be a consist set of formula on a system. We lost fact axiom, but can express more complex situation of proof from any set of  formulas. Notice that the modify of valuation does not remove the frame properties.